Add Stock Data from Database to Labels
Add Stock Data from Database to Labels in StockInfoPage & BuyPage.
Change Text Color in My Account Tab depending on Money Gaining or Losing.

+++ StockEmulator/StockEmulator/Models/AccountModel.cs
@@ -9,9 +9,9 @@ namespace StockEmulator.Models
     public class AccountModel
     {
         public string Username { get; set; }
-        public double StartingInvestment { get; set; }
+        public decimal StartingInvestment { get; set; }
         public decimal StocksValue { get; set; }
-        public double AvailableCash { get; set; }
+        public decimal AvailableCash { get; set; }
         public decimal TotalValue { get; set; }
         public decimal Position { get; set; }
         public long TotalTrans { get; set; }
